query

Module datafusion

Planner, QueryEngine implementations based on DataFusion.

Modules

error 🔒

planner 🔒

Structs

DatafusionQueryEngine

DfContextProviderAdapter

Constants

QUERY_PARALLELISM_HINT

Query parallelism hint key. This hint can be set in the query context to control the parallelism of the query execution.