query::dist_plan::analyzer

Module fallback

Fallback dist plan analyzer, which will only push down table scan node This is used when PlanRewriter produce errors when trying to rewrite the plan This is a temporary solution, and will be removed once we have a more robust plan rewriter

Structs

FallbackPlanRewriter

FallbackPlanRewriter is a plan rewriter that will only push down table scan node This is used when PlanRewriter produce errors when trying to rewrite the plan This is a temporary solution, and will be removed once we have a more robust plan rewriter It will traverse the logical plan and rewrite table scan node to merge scan node