query::dist_plan::analyzer

Module utils

Structs

AliasLayer 🔒

Functions

aliased_columns_for

Return all the aliased columns(at aliased node) for the given original columns(at original node)

aliased_columns_for_inner 🔒

get_alias_layer_from_exprs 🔒

get_alias_layer_from_node 🔒

Return a mapping of original column to all the aliased columns in current node of the plan TODO(discord9): also support merge scan node

get_alias_original_column 🔒

original_column_for

Return all the original columns(at original node) for the given aliased columns at the aliased node

original_column_for_inner 🔒

rewrite_column 🔒

rewrite_merge_sort_exprs

Rewrite the expressions of the given merge sort plan from original columns(at merge sort’s input plan) to aliased columns at the given aliased node

Type Aliases

AliasMapping

Mapping of original column in table to all the alias at current node