query::dist_plan

Module merge_sort

Merge sort logical plan for distributed query execution, roughly corresponding to the SortPreservingMergeExec operator in datafusion

Structs

MergeSortLogicalPlan

MergeSort Logical Plan, have same field as Sort, but indicate it is a merge sort, which assume each input partition is a sorted stream, and will use SortPreserveingMergeExec to merge them into a single sorted stream.

Functions

merge_sort_transformer

Turn Sort into MergeSort if possible