Module query::dist_plan::planner

[ExtensionPlanner] implementation for distributed planner

Structs

DistExtensionPlanner

MergeSortExtensionPlanner

Planner for convert merge sort logical plan to physical plan it is currently a fallback to sort, and doesn’t change the execution plan: MergeSort(MergeScan) -> Sort(MergeScan) - to physical plan -> ... It should be applied after DistExtensionPlanner

TableNameExtractor 🔒

Visitor to extract table name from logical plan (TableScan node)

Enums

TableType

Indicates the type of this table for metadata/catalog purposes.