query::optimizer

Module type_conversion

Structs

TypeConversionRule

TypeConversionRule converts some literal values in logical plan to other types according to data type of corresponding columns. Specifically:

TypeConverter 🔒

Functions

string_to_timestamp_ms 🔒

timestamp_to_timestamp_ms_expr 🔒