Function query::range_select::plan::produce_align_time

fn produce_align_time(
    align_to: i64,
    range: <TimestampMillisecondType as ArrowPrimitiveType>::Native,
    align: <TimestampMillisecondType as ArrowPrimitiveType>::Native,
    ts_column: &TimestampMillisecondArray,
    by_columns_hash: &[u64],
    modify_map: &mut HashMap<(u64, <TimestampMillisecondType as ArrowPrimitiveType>::Native), Vec<u32>>
)

Use align_to as time origin. According to align as time interval, produces aligned time. Combining the parameters related to the range query, determine for each Accumulator (hash, align_ts) define, which rows of data will be applied to it.