Module query::range_select::plan_rewrite

Macros

inconsistent_check 🔒

Structs

RangeExprRewriter

RangeExprRewriter will recursively search certain Expr, find all range_fn scalar udf contained in Expr, and collect the information required by the RangeSelect query, and finally modify the range_fn scalar udf to an ordinary column field.

RangePlanRewriter

In order to implement RangeSelect query like avg(field_0) RANGE '5m' FILL NULL, All RangeSelect query items are converted into udf scalar function in sql parse stage, with format like range_fn(avg(field_0), .....). range_fn contains all the parameters we need to execute RangeSelect. In order to correctly execute the query process of range select, we need to modify the query plan generated by datafusion. We need to recursively find the entire LogicalPlan, and find all range_fn scalar udf contained in the project plan, collecting info we need to generate RangeSelect Query LogicalPlan and rewrite th original LogicalPlan.

Functions

dispose_parse_error 🔒

have_range_in_exprs 🔒

parse_align_to 🔒

Parse the align to clause and return a UTC timestamp with unit of millisecond, which is used as the basis for dividing time slot during the align operation.

parse_duration_expr 🔒

Parse a duraion expr:

parse_expr_list 🔒

parse_expr_to_string 🔒

parse_str_expr 🔒