Module query::sql

Modules

show_create_table 🔒

Implementation of SHOW CREATE TABLE statement.

Constants

COLUMN_DEFAULT_COLUMN 🔒

COLUMN_KEY_COLUMN 🔒

COLUMN_NAME_COLUMN 🔒

COLUMN_NULLABLE_COLUMN 🔒

COLUMN_SEMANTIC_TYPE_COLUMN 🔒

COLUMN_TYPE_COLUMN 🔒

NULLABLE_NO 🔒

NULLABLE_YES 🔒

PRI_KEY 🔒

SCHEMAS_COLUMN 🔒

TABLES_COLUMN 🔒

TABLE_TYPE_COLUMN 🔒

Statics

DESCRIBE_TABLE_OUTPUT_SCHEMA 🔒

SHOW_CREATE_TABLE_OUTPUT_SCHEMA 🔒

Functions

check_file_to_table_schema_compatibility

This function checks if the column schemas from a file can be matched with the column schemas of a table.

create_table_stmt

Create a CreateTable statement from table info.

describe_column_defaults 🔒

describe_column_keys 🔒

describe_column_names 🔒

describe_column_nullables 🔒

describe_column_semantic_types 🔒

describe_column_types 🔒

describe_table

file_column_schemas_to_table

infer_file_table_schema

parse_file_table_format 🔒

prepare_file_table_files

query_from_information_schema_table 🔒

Cast a show statement execution into a query from tables in information_schema.

show_create_table

show_databases

show_tables

show_variable