Module query::window_sort

A physical plan for window sort(Which is sorting multiple sorted ranges according to input PartitionRange).

Macros

array_iter_helper 🔒

downcast_ts_array 🔒

Downcast a temporal array to a specific type

Structs

SortedRunSet 🔒

SucRun 🔒

A struct to represent a successive run in the input iterator

TimeRange 🔒

Left(start) inclusive right(end) exclusive,

WindowedSortExec

A complex stream sort execution plan which accepts a list of PartitionRange and merge sort them whenever possible, and emit the sorted result as soon as possible. This sorting plan only accept sort by ts and will not sort by other fields.

WindowedSortStream

The core logic of merging sort multiple sorted ranges

Enums

Action 🔒

Functions

check_partition_range_monotonicity 🔒

cmp_with_opts 🔒

Compare with options, note None is considered as NULL here

compute_all_working_ranges 🔒

Compute all working ranges and corresponding working sets from given overlap_counts computed from split_overlapping_ranges

find_slice_from_range 🔒

Find the slice(where start <= data < end and sort by sort_column.options) from the given range

find_successive_runs 🔒

find all successive runs in the input iterator

get_sorted_runs 🔒

return a list of non-overlapping (offset, length) which represent sorted runs, and can be used to call [DfRecordBatch::slice] to get sorted runs Returned runs will be as long as possible, and will not overlap with each other

get_timestamp_from_idx 🔒

Get timestamp from array at offset

split_batch_to_sorted_run 🔒

split batch to sorted runs

split_overlapping_ranges 🔒

return a map of non-overlapping ranges and their corresponding index (not PartitionRange.identifier but position in array) in the input PartitionRanges that is in those ranges

split_range_by 🔒

split input range by split_by range to one, two or three parts.