Crate query

This library gives various functions that are used to query a user. Each function performs the following:

Read the next command line argument, and try to parse it as an answer. If the parsing fails, panic. Otherwise, return the argument

If there are no command line arguments left, query the user for an input, and parse it as an answer. If the parsing fails, query the user again.

The “normal” usage mode is to not supply any command line arguments and just use the second functionality. However, the first is useful for testing and batch processing.

Constants

ARGV 🔒

Functions

optional

raw

vector

with_default

yes_no