query

Module dist_plan

Modules

analyzer 🔒

commutativity 🔒

merge_scan 🔒

merge_sort 🔒

Merge sort logical plan for distributed query execution, roughly corresponding to the SortPreservingMergeExec operator in datafusion

planner 🔒

[ExtensionPlanner] implementation for distributed planner

Structs

DistExtensionPlanner

DistPlannerAnalyzer

MergeScanExec

MergeScanLogicalPlan

MergeSortExtensionPlanner

Planner for convert merge sort logical plan to physical plan