query::query_engine

Trait QueryEngine

pub trait QueryEngine: Send + Sync {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn planner(&self) -> Arc<dyn LogicalPlanner>;
    fn name(&self) -> &str;
    fn describe<'life0, 'async_trait>(
        &'life0 self,
        plan: LogicalPlan,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<DescribeResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn execute<'life0, 'async_trait>(
        &'life0 self,
        plan: LogicalPlan,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_aggregate_function(&self, func: AggregateUDF);
    fn register_scalar_function(&self, func: ScalarFunctionFactory);
    fn read_table(&self, table: TableRef) -> Result<DataFrame>;
    fn engine_context(&self, query_ctx: QueryContextRef) -> QueryEngineContext;
    fn engine_state(&self) -> &QueryEngineState;
}

Required Methods

fn as_any(&self) -> &dyn Any

Returns the query engine as Any so that it can be downcast to a specific implementation.

fn planner(&self) -> Arc<dyn LogicalPlanner>

Returns the logical planner

fn name(&self) -> &str

Returns the query engine name.

fn describe<'life0, 'async_trait>( &'life0 self, plan: LogicalPlan, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<DescribeResult>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Describe the given [LogicalPlan].

fn execute<'life0, 'async_trait>( &'life0 self, plan: LogicalPlan, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Execute the given [LogicalPlan].

fn register_aggregate_function(&self, func: AggregateUDF)

Register an aggregate function.

Panics

Will panic if the function with same name is already registered.

fn register_scalar_function(&self, func: ScalarFunctionFactory)

Register a scalar function. Will override if the function with same name is already registered.

fn read_table(&self, table: TableRef) -> Result<DataFrame>

Create a DataFrame from a table.

fn engine_context(&self, query_ctx: QueryContextRef) -> QueryEngineContext

Create a QueryEngineContext.

fn engine_state(&self) -> &QueryEngineState

Retrieve the query engine state QueryEngineState