Module query::range_select::plan

Structs

RangeFirstListValue 🔒

Implementation of first_value/last_value aggregate function adapted to range query

RangeFirstListValueAcc

RangeFn

RangeFnExec 🔒

RangeSelect

RangeSelectExec

RangeSelectStream 🔒

SeriesState 🔒

Enums

ExecutionState 🔒

Fill

Functions

cast_scalar_values 🔒

linear_interpolation 🔒

use (ts[i1], data[i1]), (ts[i2], data[i2]) as endpoint, linearly interpolates element over the interval [start, end)

produce_align_time 🔒

Use align_to as time origin. According to align as time interval, produces aligned time. Combining the parameters related to the range query, determine for each Accumulator (hash, align_ts) define, which rows of data will be applied to it.

Type Aliases

Millisecond 🔒