Function query::range_select::plan_rewrite::parse_align_to

fn parse_align_to(
    args: &[Expr],
    i: usize,
    timezone: Option<&Timezone>
) -> Result<i64>

Parse the align to clause and return a UTC timestamp with unit of millisecond, which is used as the basis for dividing time slot during the align operation.

NOW: align to current execute time

Timestamp string: align to specific timestamp

An expr can be evaluated at the logical plan stage (e.g. now() - INTERVAL '1' day)

leave empty (as Default Option): align to unix epoch 0 (timezone aware)