query

Module sql

Modules

show_create_table 🔒

Implementation of SHOW CREATE TABLE statement.

Constants

COLUMN_COLLATION_COLUMN 🔒

COLUMN_COMMENT_COLUMN 🔒

COLUMN_DEFAULT_COLUMN 🔒

COLUMN_EXTRA_COLUMN 🔒

COLUMN_GREPTIME_TYPE_COLUMN 🔒

COLUMN_KEY_COLUMN 🔒

COLUMN_NAME_COLUMN 🔒

COLUMN_NULLABLE_COLUMN 🔒

COLUMN_PRIVILEGES_COLUMN 🔒

COLUMN_SEMANTIC_TYPE_COLUMN 🔒

COLUMN_TYPE_COLUMN 🔒

FIELD_COLUMN 🔒

FLOWS_COLUMN 🔒

INDEX_CARDINALITY_COLUMN 🔒

INDEX_COLUMN_NAME_COLUMN 🔒

INDEX_COMMENT_COLUMN 🔒

INDEX_EXPRESSION_COLUMN 🔒

INDEX_INDEX_TYPE_COLUMN 🔒

INDEX_KEY_NAME_COLUMN 🔒

INDEX_NONT_UNIQUE_COLUMN 🔒

INDEX_PACKED_COLUMN 🔒

INDEX_SEQ_IN_INDEX_COLUMN 🔒

INDEX_SUB_PART_COLUMN 🔒

INDEX_TABLE_COLUMN 🔒

SHOW index columns

INDEX_VISIBLE_COLUMN 🔒

NO_STR 🔒

OPTIONS_COLUMN 🔒

PRI_KEY 🔒

SCHEMAS_COLUMN 🔒

TABLE_TYPE_COLUMN 🔒

TIME_INDEX 🔒

VIEWS_COLUMN 🔒

YES_STR 🔒

Statics

DESCRIBE_TABLE_OUTPUT_SCHEMA 🔒

SHOW_CREATE_DATABASE_OUTPUT_SCHEMA 🔒

SHOW_CREATE_FLOW_OUTPUT_SCHEMA 🔒

SHOW_CREATE_TABLE_OUTPUT_SCHEMA 🔒

SHOW_CREATE_VIEW_OUTPUT_SCHEMA 🔒

Functions

check_file_to_table_schema_compatibility

This function checks if the column schemas from a file can be matched with the column schemas of a table.

create_table_stmt

Create a CreateTable statement from table info.

describe_column_defaults 🔒

describe_column_keys 🔒

describe_column_names 🔒

describe_column_nullables 🔒

describe_column_semantic_types 🔒

describe_column_types 🔒

describe_table

file_column_schemas_to_table

infer_file_table_schema

null 🔒

parse_file_table_format 🔒

prepare_file_table_files

query_from_information_schema_table 🔒

Cast a show statement execution into a query from tables in information_schema.

replace_column_in_expr 🔒

Replaces column identifier references in a SQL expression. Used for backward compatibility where old column names should work with new ones.

show_charsets

Execute SHOW CHARSET statement and returns the Output if success.

show_collations

Execute SHOW COLLATION statement and returns the Output if success.

show_columns

Execute SHOW COLUMNS statement.

show_create_database

show_create_flow

show_create_foreign_table_for_pg

show_create_table

show_create_view

show_databases

show_flows

Execute [ShowFlows] statement and return the [Output] if success.

show_index

Execute SHOW INDEX statement.

show_processlist

show_region

Execute SHOW REGION statement.

show_search_path

show_status

show_table_status

Execute [ShowTableStatus] statement and return the [Output] if success.

show_tables

Execute [ShowTables] statement and return the [Output] if success.

show_variable

show_views

Execute [ShowViews] statement and return the [Output] if success.